Broke up huge files, places events and groups into directories
This way its a little easier to navigate the code.
